refact: use base SymbolID for SymbolVAR
