+++ src/api/constants.py
@@ -47,7 +47,7 @@ class CLASS(str, enum.Enum):
 
     @classproperty
     def classes(cls):
-        return (cls.unknown, cls.var, cls.array, cls.function, cls.sub, cls.const, cls.label)
+        return cls.unknown, cls.var, cls.array, cls.function, cls.sub, cls.const, cls.label
 
     @classmethod
     def is_valid(cls, class_: Union[str, "CLASS"]):
